+++ CHANGELOG.md
@@ -16,12 +16,17 @@ CHANGELOG
 - Double click edits the bot or the queue
 - Code cleanup: New modal window for performing multiple dialog actions. Intercepting visjs internals without using external Hammer library.
 - Enter key submits the edit bot dialog.
+- Clicking on the side-menu bot name cycles amongst bot instances in the canvas.
+- Got rid of the dumb "#load" and "#new" parameter → "#new" converted into a standard URL parameter "configuration=new".
+- Button to turn on/off physics
 
 #### Management
 
 #### Monitor
 * Fix link to monitor page (#157).
 * The new parameters panel that displays the bot's parameters on the bot's monitor page. For parameters which are filenames, the content of these files will be shown as text if they reside in the allowed paths (set in config.php).
+* When needed, Path column appears, resizing the Queues panel accordingly.
+* Connected with the configuration panel → a link appears next to the bot title that takes you directly to the bot
 
 #### Check